Implement workaround
