+++ packages/devtools_app/lib/src/shared/utils/_focus_utils_desktop.dart
@@ -2,6 +2,10 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file or at https://developers.google.com/open-source/licenses/bsd.
 
-void reloadIframe() {
+void addBlurListener() {
+  // No-op for desktop platforms.
+}
+
+void removeBlurListener() {
   // No-op for desktop platforms.
 }
