+++ packages/devtools_app/lib/src/standalone_ui/ide_shared/property_editor/utils/_utils_web.dart
@@ -2,23 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file or at https://developers.google.com/open-source/licenses/bsd.
 
-import 'dart:js_interop';
-
 import 'package:web/web.dart';
 
 void reloadIframe() {
   window.location.reload();
-}
-
-void addBlurListener() {
-  window.addEventListener('blur', _onBlur.toJS);
-}
-
-void removeBlurListener() {
-  window.removeEventListener('blur', _onBlur.toJS);
-}
-
-void _onBlur(Event _) {
-  final inputElement = document.activeElement as HTMLElement?;
-  inputElement?.blur();
 }
